+++ modules/gui/src/com/haulmont/charts/gui/amcharts/model/charts/CoordinateChart.java
@@ -34,7 +34,7 @@ public abstract class CoordinateChart<T extends CoordinateChart> extends Abstrac
 
     private Double startAlpha;
 
-    private Integer startDuration;
+    private Double startDuration;
 
     private AnimationEffect startEffect;
 
@@ -102,12 +102,12 @@ public abstract class CoordinateChart<T extends CoordinateChart> extends Abstrac
     }
 
     @Override
-    public Integer getStartDuration() {
+    public Double getStartDuration() {
         return startDuration;
     }
 
     @Override
-    public T setStartDuration(Integer startDuration) {
+    public T setStartDuration(Double startDuration) {
         this.startDuration = startDuration;
         return (T) this;
     }
